開発者用画面Menu表示View共通化
