+++ EngineerMemo/Presentation/Screen/Debug/CoreData/Create/Memo/DebugMemoCreateContentView.swift
@@ -51,7 +51,7 @@
             backgroundColor = .primary
 
             addSubview(body) {
-                $0.edges.equalToSuperview()
+                $0.top.leading.trailing.equalToSuperview()
             }
         }
     }
