+++ EngineerMemo/Presentation/Screen/Debug/CoreData/Create/Profile/DebugProfileCreateContentView.swift
@@ -81,7 +81,7 @@
             backgroundColor = .primary
 
             addSubview(body) {
-                $0.edges.equalToSuperview()
+                $0.top.leading.trailing.equalToSuperview()
             }
         }
     }
